自动打分

+++ src/main/java/com/hw/mapper/SequenceMapper.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Param;
 
 
 public interface SequenceMapper {
-     Long  getNextSequence(@Param("sequence") String  sequence);
+     Integer  getNextSequence(@Param("sequence") String  sequence);
 }
